# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Camila SaaS - API para emprendedores locales
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerStore
      tags: [auth]
      summary: Register a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterStoreRequest"
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Store already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/stores:
    get:
      operationId: adminGetAllStores
      tags: [admin]
      summary: Get all stores (superadmin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [trial, active, expired, suspended]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of stores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStores"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/stores/{storeId}:
    get:
      operationId: adminGetStore
      tags: [admin]
      summary: Get a specific store (superadmin)
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"

    patch:
      operationId: adminUpdateStore
      tags: [admin]
      summary: Update store (superadmin)
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateStoreRequest"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /admin/stores/{storeId}/license:
    patch:
      operationId: adminUpdateLicense
      tags: [admin]
      summary: Update store license (superadmin)
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLicenseRequest"
      responses:
        "200":
          description: License updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/License"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        store:
          $ref: "#/components/schemas/Store"
        token:
          type: string
      required:
        - user
        - store
        - token
    CurrentUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        store:
          $ref: "#/components/schemas/Store"
        license:
          $ref: "#/components/schemas/License"
      required:
        - user
        - store
    RegisterStoreRequest:
      type: object
      properties:
        businessName:
          type: string
        businessType:
          type: string
          enum: [clothing, restaurant, bakery, fair_booth, general_catalog]
        documentType:
          type: string
          enum: [DNI, RUC10, RUC20]
        documentNumber:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        address:
          type: string
        district:
          type: string
        logoUrl:
          type: string
      required:
        - businessName
        - businessType
        - documentType
        - documentNumber
        - ownerName
        - phone
        - email
        - password
        - district
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
      required:
        - token
        - password
...
    AdminUpdateStoreRequest:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: string
        district:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
...
    UpdateLicenseRequest:
      type: object
      properties:
        status:
          type: string
          enum: [trial, active, expired, suspended]
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - status
    User:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, store_admin, store_staff, cashier]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - name
        - email
        - role
        - isActive
        - createdAt
    Store:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        businessType:
          type: string
          enum: [clothing, restaurant, bakery, fair_booth, general_catalog]
        documentType:
          type: string
          enum: [DNI, RUC10, RUC20]
        documentNumber:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        district:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        primaryColor:
          type: string
        description:
          type: string
        whatsapp:
          type: string
        socialInstagram:
          type: string
        socialFacebook:
          type: string
        isActive:
          type: boolean
        license:
          $ref: "#/components/schemas/License"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessName
        - businessType
        - documentType
        - documentNumber
        - ownerName
        - phone
        - email
        - district
        - isActive
        - createdAt
    License:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        status:
          type: string
          enum: [trial, active, expired, suspended]
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        notes:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - status
        - updatedAt
    PaginatedStores:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
...