# lib/api-zod/src/generated/types/adminUpdateStoreRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Camila SaaS - API para emprendedores locales
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUpdateStoreRequest {
  businessName?: string;
  address?: string;
  district?: string;
  phone?: string;
  isActive?: boolean;
}